Change calculation flow to as numbers are entered
Changed when calculations are performed to conform to The Odin Project
guidelines. Now as the user inputs numbers the total will be displayed
after they select operation key for the next calculation.
This changed meant an array of numbers and operators were unnecessary
and the associated functions and variables have been removed.